+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.0] - 2018-09-10
+
+### Added
+- Slim Middleware for HTTP query conversion. (AidanCasey\HttpParser\Middleware\Slim)
+
+### Changed
+- HttpToZend::ConvertHTTPRequest to HttpToZend::ConvertHttpRequest.
+
 ## [1.1.0] - 2018-09-04
 
 ### Changed
@@ -21,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HttpToZend class for conversion of request parameters to database statements.
 - UnsupportedRequestMethod for any unsupported request types.
 
-[Unreleased]: https://github.com/aidan-casey/http-parser/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/aidan-casey/http-parser/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/aidan-casey/http-parser/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/aidan-casey/http-parser/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/aidan-casey/http-parser/releases/tag/v1.0.0